# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "代理管理服务",
        "title": "代理管理服务",
        "contact": {},
        "version": "V0.1"
    },
    "paths": {
...
        "/api/v1/menu": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取菜单列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "获取菜单列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "条件",
                        "name": "condition",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页数",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "总数  由服务器返回回去",
                        "name": "total",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code: 200 成功",
                        "schema": {
                            "$ref": "#/definitions/controller.Full"
                        }
                    },
                    "500": {
                        "description": "错误返回内容",
                        "schema": {
                            "$ref": "#/definitions/controller.Base"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "添加菜单信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "添加菜单信息",
                "parameters": [
                    {
                        "description": "添加信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/repository.CreateMenu"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code: 200 成功",
                        "schema": {
                            "$ref": "#/definitions/controller.Base"
                        }
                    },
                    "500": {
                        "description": "错误返回内容",
                        "schema": {
                            "$ref": "#/definitions/controller.Base"
                        }
                    }
                }
            }
        },
        "/api/v1/menu/state/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "修改菜单状态",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "修改菜单状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code: 200 成功",
                        "schema": {
                            "$ref": "#/definitions/controller.Base"
                        }
                    },
                    "500": {
                        "description": "错误返回内容",
                        "schema": {
                            "$ref": "#/definitions/controller.Base"
                        }
                    }
                }
            }
        },
        "/api/v1/menu/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "通过编码获取菜单",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "通过编码获取菜单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code: 200 成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controller.Full"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Menu"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "错误返回内容",
                        "schema": {
                            "$ref": "#/definitions/controller.Base"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "修改菜单信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "修改菜单信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "修改信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/repository.ModifyMenu"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code: 200 成功",
                        "schema": {
                            "$ref": "#/definitions/controller.Base"
                        }
                    },
                    "500": {
                        "description": "错误返回内容",
                        "schema": {
                            "$ref": "#/definitions/controller.Base"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除菜单信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "菜单管理"
                ],
                "summary": "删除菜单信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code: 200 成功",
                        "schema": {
                            "$ref": "#/definitions/controller.Base"
                        }
                    },
                    "500": {
                        "description": "错误返回内容",
                        "schema": {
                            "$ref": "#/definitions/controller.Base"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.Base": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "message": {
                    "description": "状态消息",
                    "type": "string"
                },
                "requestDatetime": {
                    "description": "请求时间",
                    "type": "string"
                },
                "requestId": {
                    "description": "请求ID",
                    "type": "string"
                },
                "responseDatetime": {
                    "description": "返回时间",
                    "type": "string"
                }
            }
        },
        "controller.Full": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "返回数据"
                },
                "message": {
                    "description": "状态消息",
                    "type": "string"
                },
                "requestDatetime": {
                    "description": "请求时间",
                    "type": "string"
                },
                "requestId": {
                    "description": "请求ID",
                    "type": "string"
                },
                "responseDatetime": {
                    "description": "返回时间",
                    "type": "string"
                }
            }
        },
...
        "model.Menu": {
            "type": "object",
            "required": [
                "name",
                "parentId",
                "route",
                "title"
            ],
            "properties": {
                "auth": {
                    "description": "菜单权限级别",
                    "type": "integer"
                },
                "createdAt": {
                    "description": "创建时间",
                    "type": "string"
                },
                "icon": {
                    "description": "菜单图标",
                    "type": "string"
                },
                "id": {
                    "description": "编码",
                    "type": "string"
                },
                "level": {
                    "description": "菜单层级级别 0 代表第一层",
                    "type": "integer"
                },
                "name": {
                    "description": "菜单名称",
                    "type": "string"
                },
                "parentId": {
                    "description": "父级菜单编码 如果没有则传\"0\"",
                    "type": "string"
                },
                "route": {
                    "description": "路由",
                    "type": "string"
                },
                "state": {
                    "description": "状态 0 停用 1 启用",
                    "type": "integer"
                },
                "title": {
                    "description": "菜单标题",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "修改时间",
                    "type": "string"
                }
            }
        },
...
        "repository.CreateMenu": {
            "type": "object",
            "required": [
                "name",
                "parentId",
                "route",
                "title"
            ],
            "properties": {
                "auth": {
                    "description": "菜单权限级别",
                    "type": "integer"
                },
                "icon": {
                    "description": "菜单图标",
                    "type": "string"
                },
                "level": {
                    "description": "菜单层级级别 0 代表第一层",
                    "type": "integer"
                },
                "name": {
                    "description": "菜单名称",
                    "type": "string"
                },
                "parentId": {
                    "description": "父级菜单编码 如果没有则传\"0\"",
                    "type": "string"
                },
                "route": {
                    "description": "路由",
                    "type": "string"
                },
                "state": {
                    "description": "状态 0 停用 1 启用",
                    "type": "integer"
                },
                "title": {
                    "description": "菜单标题",
                    "type": "string"
                }
            }
        },
...
        "repository.ModifyMenu": {
            "type": "object",
            "required": [
                "name",
                "parentId",
                "route",
                "title"
            ],
            "properties": {
                "auth": {
                    "description": "菜单权限级别",
                    "type": "integer"
                },
                "icon": {
                    "description": "菜单图标",
                    "type": "string"
                },
                "id": {
                    "description": "编码",
                    "type": "string"
                },
                "level": {
                    "description": "菜单层级级别 0 代表第一层",
                    "type": "integer"
                },
                "name": {
                    "description": "菜单名称",
                    "type": "string"
                },
                "parentId": {
                    "description": "父级菜单编码 如果没有则传\"0\"",
                    "type": "string"
                },
                "route": {
                    "description": "路由",
                    "type": "string"
                },
                "state": {
                    "description": "状态 0 停用 1 启用",
                    "type": "integer"
                },
                "title": {
                    "description": "菜单标题",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}